refactor(updater): updater_core.py(770줄)를 책임별 5개 모듈로 분리 (v2.4.26)
이슈 #87 Phase 5에서 후속으로 미뤄둔 분리를 완료한다. 770줄 단일 파일을
환경 감지 / 3개 백엔드 / 실행 프레임워크로 나누고, 기존 호출부가 코드
변경 없이 동작하도록 updater_core를 thin re-export shim으로 둔다.

* updater/environment.py — 런타임 감지
  - _is_standalone_build / is_velopack_environment / get_velopack_update_exe
  - is_pip_environment
* updater/velopack.py — Velopack 다운로드/적용
  - VelopackUpdater + VelopackDownloadError
* updater/pip_updater.py — pip-install 환경 업그레이드
  - PipUpdater
* updater/legacy.py — 직접 다운로드/실행 경로
  - LegacyInstallerUpdater + Updater(레거시 호환) + GITHUB_RELEASES_URL
* updater/executors.py — 비-GUI 실행 프레임워크
  - UpdateExecutor ABC + PipExecutor / VelopackExecutor / LegacyExecutor
  - create_update_executor / get_updater 팩토리
* updater/updater_core.py — 18개 공개 심볼 re-export shim

테스트 패치 대상 정정:
  tests/test_velopack_updater.py / tests/test_update_executors.py가
  `mock.patch.object(updater_core, "X", …)` 형태로 shim의 네임스페이스를
  패치했지만, `from X import Y`는 importer 측에 Y를 바인딩하므로 실제
  호출부(velopack.py / executors.py)의 Y는 patch에 영향받지 않는 상태였다.
  이를 새 모듈을 가리키도록 정정 — `velopack_module.subprocess`,
  `velopack_module.sys`, `velopack_module.get_velopack_update_exe`,
  `executors_module.GITHUB_RELEASES_URL`. 16건 테스트 모두 통과.

릴리스 노트:
  이 커밋과 함께 v2.4.26으로 bump하며, CHANGELOG에는 본 분리 외에도
  같은 PR에 포함된 두 항목(Nuitka --include-module 50+→8 정리, CI 워크
  플로 3곳의 인라인 Nuitka 명령을 build_nuitka.py 단일 호출로 통합)을
  같이 기록한다.
